Create post to dumpsite circle, bypass feed
This code sets up the Twitter API and defines a function called sendDumpsiteTweet(lat: Double, long: Double) that allows a user to send a tweet with the location of a dumpsite, and also processes photos. The tweet is sent to the DumpsiteData Circle, which collects similar data from other users. The code uses the twitter4j library to interact with the Twitter API and create the tweets.

Code contains three options, with Option C preferred:

This updated version of the sendDumpsiteTweet(lat: Double, long: Double, photo: Option[String]) function uses a List[String] instead of an Option[String] to allow, but not require, multiple photos to be uploaded. The photos parameter can either be an empty List[String] (if no photos are provided) or a List[String] containing one or more URLs of photos. The code uploads each photo to Twitter's servers if it is provided, and includes all of the uploaded photos in the tweet that is sent to the DumpsiteData Circle.