Create DumpsiteDataMap (google)
The showDumpsiteDataOnMap(map: Map, dumpsiteData: List[DumpsiteData]) function adds markers to the map at the locations of the Dumpsites and creates an overlay that shows the Dumpsite data on the map. To do list: use the map and dumpsiteData parameters to add markers to the map at the locations of the Dumpsites and to create an overlay that shows the Dumpsite data on the map.

The createDumpsiteDataMap() function creates the map and calls the showDumpsiteDataOnMap(map: Map, dumpsiteData: List[DumpsiteData]) function to show the Dumpsite data on the map. This function returns the map for display in TwitterCA and in Twitter native.

The showDumpsiteDataInTweet(tweet: Tweet, map: Map) function displays the map in a Tweet, allowing users to view the Dumpsite data on a map directly in Twitter. To do: use the tweet and map parameters to create a new Tweet that includes the map and then send the Tweet using the Twitter API.