Create add heatmap overlay
This code creates a HeatmapTileProvider that generates the heatmap data from the DumpsiteData Circle data, using the getDumpsiteTweets() function to retrieve the latitude and longitude information for each Dumpsite. It then creates a HeatmapOverlay using the HeatmapTileProvider and adds it to the map using the map.addOverlay() method. This will add a heatmap layer to the map that shows the intensity of the Dumpsites on the map.