Create scoreboard
This code uses the Twitter API to search for tweets with the #CleanApp hashtag and count the number of tweets that the user has sent. It then updates the user's profile with their total point value, using the Twitter.updateProfile(profile: Profile) method. Finally, it creates a global leaderboard that displays the top users based on their point value, using the Twitter.search(query: Query) method to search for tweets with the #CleanApp leaderboard hashtag and iterating through the tweets to display the user's name and total point value.